requirements.txt
absl-py==1.2.0
alabaster==0.7.8
apturl==0.5.2
astropy==5.1
asttokens==2.0.8
astunparse==1.6.3
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
capstone==5.0.0rc2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
colored-traceback==0.3.0
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
executing==1.1.0
fasteners==0.14.1
flatbuffers==22.9.24
fonttools==4.37.1
future==0.18.2
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==5.0.0
intervaltree==3.1.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==2.10.1
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.2
matplotlib-inline==0.1.6
monotonic==1.5
nest-asyncio==1.5.6
netifaces==0.10.4
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
paramiko==2.6.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plumbum==1.8.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
psycopg2-binary==2.9.4
pure-eval==0.2.2
pwntools==4.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodome==3.15.0
pycups==1.9.73
pyelftools==0.29
pyerfa==2.0.0.1
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-resize-image==1.1.20
pytz==2022.4
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
ROPGadget==7.1
rpyc==5.2.3
rsa==4.9
scipy==1.9.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==1.8.5
sphinx-rtd-theme==1.0.0
stack-data==0.5.1
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicorn==2.0.1.post1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
zipp==3.8.1
